+++ src/components/Settings/tabs/Behavior.tsx
@@ -1,4 +1,4 @@
-import { Stack, Switch } from "@mantine/core";
+import { Kbd, Stack, Switch } from "@mantine/core";
 import { adapter } from "~/adapter";
 import { useStable } from "~/hooks/stable";
 import { actions, store } from "~/store";
@@ -27,6 +27,11 @@ export function GeneralTab({ config }: GeneralTabProps) {
 		updateConfig();
 	});
 
+	const togglePinned = useStable((e: React.ChangeEvent<HTMLInputElement>) => {
+		store.dispatch(actions.setWindowPinned(e.target.checked));
+		updateConfig();
+	});
+
 	return (
 		<Stack spacing="xs">
 			{adapter.isUpdateCheckSupported && (
@@ -42,6 +47,13 @@ export function GeneralTab({ config }: GeneralTabProps) {
 			<Setting label="Query error checking">
 				<Switch checked={config.errorChecking} onChange={setErrorChecking} />
 			</Setting>
+			
+
+			{adapter.isPinningSupported && (
+				<Setting label={<>Window always on top <Kbd size="xs">F11</Kbd></>}>
+					<Switch checked={config.isPinned} onChange={togglePinned} />
+				</Setting>
+			)}
 		</Stack>
 	);
 }