+++ src/typings/surreal.d.ts
@@ -35,6 +35,5 @@ export interface SurrealInfoTB {
 	events: Record<string, string>;
 	fields: Record<string, string>;
 	indexes: Record<string, string>;
-	analyzers: Record<string, string>;
 	tables: Record<string, string>;
 }