+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Designer UI improvements
 	- Field names and types are now truncated when they are too long
 	- Limit tables to displaying 7 fields, with a button to view more
+- Replaced the window pinning toolbar button with a settings entry and hotkey (F11)
 - Fixed web app always starting in light theme for a split second
 
 ## 1.9.2 - SurrealDB 1.0 support